% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extrct_vssl}
\alias{extrct_vssl}
\title{Prefix Bundle Extraction Vessel}
\usage{
extrct_vssl(x, num, srch_patt)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}

\item{num}{a number, as numeric or character, that matches the first
occurrence of a number in the prefix bundle. Determined automatically in
\code{gather_n_move_prefix_num_bundle()}}

\item{srch_patt}{The regular expression search pattern, expressed as a string}
}
\description{
A helper function used to simplify the extraction of certain prefix bundles, used in
\code{\link{gather_n_move_prefix_num_bundle}}.
}
\examples{
extrct_vssl(x = "1a. How Was Brunch?", 1, "[^[:alnum:]]([0-9]){1,3}")
}
\keyword{internal}
