% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_n_move_prefix_num_bundle}
\alias{gather_n_move_prefix_num_bundle}
\title{Extract 'prefix bundle'}
\usage{
gather_n_move_prefix_num_bundle(x, relo_2_end = T, sep = "_")
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}

\item{relo_2_end}{logical, if \code{TRUE} then the prefix bundle, if it exists
will be scooted to the end of the string x, else it will receive the
necessary "_" prefix}

\item{sep}{string of only one character long, intended to separate words from
one another, not a starting character if original term has a numeric prefix!}
}
\description{
A 'prefix bundle' is a string that starts with a numeric digit and is
followed by 1 or 2 alpha characters. The string is passed to internal
\code{\link{read_words}} function to parse out 'bundle' alphas from
'non-bundle' alphas found in the term. Returns the "bundle prefix" with no
underscore prefix.
}
\details{
Currently, this function accounts for the following types of prefix
bundles, where 3 is used as an example starting digit:
\itemize{
\item x = "3a. hey" will return "3a"
\item x = "3_17a_hey" will return "3_17a"
\item x = "3_17_hey" will return "3_17"
\item x = "3_17hey" will return "3_17"
\item x = "3hey" will return "3"
}
}
\examples{
gather_n_move_prefix_num_bundle(x = "iAteABunch_of_grapesUntil99\%Full")
gather_n_move_prefix_num_bundle(x = "1a. How Was Brunch?")
gather_n_move_prefix_num_bundle(x = "2aa How Was Brunch2day?")
gather_n_move_prefix_num_bundle(x = "30abHow Was Brunch2day?")
gather_n_move_prefix_num_bundle(x = c("30ab.How Was Brunch2day?", "40AHow Was Brunch2day?"))
}
\keyword{internal}
