% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefix_num}
\alias{prefix_num}
\title{Extract Starting Prefix number}
\usage{
prefix_num(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Return number if a string starts with a number, "" otherwise
}
\examples{
if(interactive()) {
  prefix_num(x = "iAteABunch_of_grapesUntil99\%Full")
  prefix_num(x = "1a.How Was Brunch? Still $2.99?")
}
}
\keyword{internal}
