% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_words_nums_no_sym}
\alias{read_words_nums_no_sym}
\title{Uses a smattering of string functions to separate words and/or numbers that
may be smashed together so that the the character element is more human
readable}
\usage{
read_words_nums_no_sym(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Uses a smattering of string functions to separate words and/or numbers that
may be smashed together so that the the character element is more human
readable
}
\examples{
read_words_nums_no_sym(x = "iAteABunch_of_grapesUntil99\%Full")
}
\keyword{internal}
