% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_sym}
\alias{replace_sym}
\title{Find and replace special characters or symbols}
\usage{
replace_sym(x, replace = " ")
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}

\item{replace}{character to replace the special characters or symbols with;
Defaults to blank space.}
}
\description{
Find and replace special characters or symbols
}
\examples{
replace_sym("i.Ate-1grape?")
}
\keyword{internal}
