% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{starts_with_number}
\alias{starts_with_number}
\title{Indicate 1 if a string starts with a number, 0 otherwise}
\usage{
starts_with_number(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Indicate 1 if a string starts with a number, 0 otherwise
}
\examples{
if(interactive()) {
  starts_with_number("iAteABunch_of_grapesUntil99\%Full")
  starts_with_number("1a.How Was Brunch?")
}
}
\keyword{internal}
