% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyRename-package.R
\docType{package}
\name{tidyRename-package}
\alias{tidyRename}
\alias{tidyRename-package}
\title{tidyRename: Automagically rename wordy variables}
\description{
Reduces by automating the creation of submission compliant variable attributes (i.e. <=8 characters for variable names, <=40 characters for labels). Concepts can be expanded to CSR activities i.e., SDTM variables mapping process etc. to ensure uniformity among users and eliminates wasted time manually creating easily-readable variable names for high dimensional data sets.
}
\author{
\strong{Maintainer}: Aaron Clark \email{clark.aaronchris@gmail.com} (\href{https://orcid.org/0000-0002-0123-0970}{ORCID})

}
