% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_names_log}
\alias{var_names_log}
\title{var_names_log}
\usage{
var_names_log(tidy_names_df, verbose)
}
\arguments{
\item{tidy_names_df}{placeholder}

\item{verbose}{placeholder}
}
\description{
Functions to output user messages, usually relating to differences
found between .df and the metacore object
}
\keyword{internal}
