% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{words_}
\alias{words_}
\title{Find and replace _'s (underscores) with a blank space " "}
\usage{
words_(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Find and replace _'s (underscores) with a blank space " "
}
\examples{
if(interactive()) {
  words_("I_Ate")
}
}
\keyword{internal}
