% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{words_aB}
\alias{words_aB}
\title{Insert a space between a lowercase character and an uppercase to
separate a character string into assumed 'words'}
\usage{
words_aB(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Insert a space between a lowercase character and an uppercase to
separate a character string into assumed 'words'
}
\examples{
words_aB("iAte")
}
\keyword{internal}
