% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{words_ABb}
\alias{words_ABb}
\title{Find when there are two capital letters next to each other, followed by a
lower case letter, then add a space between the two capital letters to
separate the assumed 'words'}
\usage{
words_ABb(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Find when there are two capital letters next to each other, followed by a
lower case letter, then add a space between the two capital letters to
separate the assumed 'words'
}
\examples{
words_ABb("IAte")
}
\keyword{internal}
