% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{words_grapes}
\alias{words_grapes}
\title{Find and replace Percent symbols with "Pct"}
\usage{
words_grapes(x)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}
}
\description{
Find and replace Percent symbols with "Pct"
}
\examples{
words_grapes("Johnny owes me 90\% of his sandwich")
}
\keyword{internal}
