% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xpt_validate_var_names}
\alias{xpt_validate_var_names}
\title{Execute checks on variable names}
\usage{
xpt_validate_var_names(varnames, list_vars_first = TRUE, err_cnd = character())
}
\arguments{
\item{varnames}{a vector of variable namess}

\item{list_vars_first}{a logical}

\item{err_cnd}{character string, either empty of containing an error message}
}
\description{
Execute checks on variable names
}
\keyword{internal}
